Implement 'by index' functions (#37)
* Implement 'by index' functions

Implemented H5Gget_info_by_name, H5Oget_info_by_idx, H5Aget_info_by_idx,
H5Aget_name_by_idx, and H5Aopen_by_idx.

Also patched a memory error in H5Gget_info_by_name, and
cleaned up some things that were causing warnings on build.

* Enable additional H5Aopen* calls in test_rest_vol

A couple calls to H5Aopen_by_idx in test_rest_vol.c provided an
erroneous relative path, so I replaced those paths with "."

Specifically, the erroneous calls provided a datatype as a
parent object, then provided the name of that datatype as a relative
path.  Since the datatype doesn't contain itself, this would always
lead to the object not being found.

I also modified H5Aopen_by_idx to not make an unnecessary request to
the server when the relative path is ".", because then the provided
parent object is the same as the target object, and no path traversal
is needed.

+++ examples/rv_compound.c
@@ -67,7 +67,7 @@ main(void)
      */
     for (i = 0; i < LENGTH; i++) {
         s1[i].a = i;
-        s1[i].b = i * i;
+        s1[i].b = (float)(i * i);
         s1[i].c = 1. / (i + 1);
     }
